requirements.txt
Django~=2.2.4
django-model-utils~=4.0.0
djangorestframework~=3.11.1
django-rest-knox~=4.1.0